source/requirements.txt
simplematrixbotlib
pyyaml
Mastodon.py
bs4
Pillow